requirements.txt
pyyaml==5.4.1
prometheus-client==0.10.1